Move logic to SimpleSearch.logic.ts
