+++ frontend/src/services/searchService.unit.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, type Mock } from "vitest";
 import { axiosInstance } from "./httpClient";
-import { advancedSearch, search } from "./searchService";
+import { advancedSearch } from "./searchService";
 import { mockNuxtImport } from "@nuxt/test-utils/runtime";
 import { DocumentKind } from "~/types";
 
@@ -23,54 +23,6 @@ mockNuxtImport("useRuntimeConfig", () => {
 });
 
 describe("searchService", () => {
-  describe("search", () => {
-    it("should call axios.get with the correct URL and parameters", async () => {
-      // Arrange
-      const params = {
-        query: "test",
-        itemsPerPage: 10,
-        pageNumber: 1,
-        sort: "date",
-        category: "R.Urteil",
-        date: "2023-01-01",
-        court: "court1",
-        temporalCoverage: "0000",
-      };
-
-      const expectedParams = new URLSearchParams({
-        searchTerm: "test",
-        size: "10",
-        pageIndex: "1",
-        sort: "date",
-        dateFrom: "2023-01-01",
-        dateTo: "2023-01-01",
-        typeGroup: "Urteil",
-        court: "court1",
-        temporalCoverageFrom: "0000",
-        temporalCoverageTo: "0000",
-      });
-      const expectedURL = "/v1/case-law?" + expectedParams;
-      const expectedConfig = {
-        timeout: 10000,
-        headers: {
-          Accept: "application/json",
-        },
-        baseURL: "https://backend",
-      };
-
-      (axiosInstance.get as Mock).mockResolvedValue({ data: {} });
-
-      // Act
-      await search(params);
-
-      // Assert
-      expect(axiosInstance.get).toHaveBeenCalledWith(
-        expectedURL,
-        expectedConfig,
-      );
-    });
-  });
-
   describe("advancedSearch", () => {
     it("should call axios.get with the correct URL and parameters", async () => {
       // Arrange
